+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Mission Control location list shows a count of related events/community experiences (#322)
 - Add Google Analytics Event when participant joins (#324)
 - Survey linking for all participants and first time participants
+- Participant Count and Minutes by Program Report
+- First time users to Activity by Program Report
 
 ### Changed
 - Mission Control events navbar icon is now a drop-down (#322)